Moving files around to amek more sense; adding context menu
