+++ x-pack/plugins/infra/public/components/waffle/lib/apply_wafflemap_layout.ts
@@ -7,7 +7,7 @@ import { first, sortBy } from 'lodash';
 import {
   isWaffleMapGroupWithGroups,
   isWaffleMapGroupWithNodes,
-} from '../../../containers/libs/type_guards';
+} from '../../../containers/map/type_guards';
 import { InfraWaffleMapGroup } from '../../../lib/lib';
 import { sizeOfSquares } from './size_of_squares';
