+++ x-pack/plugins/infra/public/containers/map/nodes_to_wafflemap.ts
