+++ x-pack/plugins/infra/public/containers/map/type_guards.ts
